+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#60](https://github.com/seznam/slo-exporter/pull/60) New module eventMetadataRenamer
 
 ### Changed
 - [#54](https://github.com/seznam/slo-exporter/pull/54) Use strict unmarshal in relabel module